+++ apps/desktop/src/lib/components/common/Icon.svelte
@@ -62,6 +62,7 @@
 		'sliders-horizontal':
 			'<path d="M10 5H3" /><path d="M12 19H3" /><path d="M14 3v4" /><path d="M16 17v4" /><path d="M21 12h-9" /><path d="M21 19h-5" /><path d="M21 5h-7" /><path d="M8 10v4" /><path d="M8 12H3" />',
 		plus: 'M12 4v16m8-8H4',
+		minus: 'M20 12H4',
 		grid: 'M4 6h16M4 10h16M4 14h16M4 18h16',
 		terminal: 'M8 9l3 3-3 3m5 0h3M5 20h14a2 2 0 002-2V6a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z',
 		tag: 'M7 7h.01M7 3h5c.512 0 1.024.195 1.414.586l7 7a2 2 0 010 2.828l-7 7a2 2 0 01-2.828 0l-7-7A2 2 0 013 12V7a4 4 0 014-4z',